Removed immediate_mode as a trigger for updates.
The immediate_mode originated in the Blynclight support where I
wanted to update the light as soon as field changed without calling
an "update/write" function. This model translated poorly to other
vendor's lights, so it made more sense to remove the immediate update
(and all it's complexities!) and replace it with the inverse-type
operation in the form of a context manager that initiates I/O to
the device when the context manager exits. This serves as a visual
grouping of updates to the in-memory state of a device instead of
just terminating it all with a `write` call.