Raising USBLightIOError from write methods
The hid.device.write and hid.device.send_feature_report IO
methods return -1 when the methods encounter an error (usually
the device has been unplugged and the USB system disavows all
knowledge of the device). Started checking the return value of
HID IO methods and raising USBLightIOError to signal up the
stack that there's a problem.