Possible fix for #14 Multiple of the Same Device
The implementation of `busylight.lights.USBLight` took the easy way
out of opening the underlying USB human interface device (HID), using
the vendor identifier and product identifier for each device. If there
are multiple devices with the same vendor and product identfiers, the
first open attempt will open the first device in the USB
chain. Subsequent opens will fail since the first device is matched by
`hid.device().open` and it's already open.

To fix this, I've modified `USBLight.__init__` (and sub-classes) to
take a third argument `path: bytes` which is found in the `dict`
obtained from `hid.enumerate`. This new `path` argument (and class
property) is used to open the HID using `hid.device().open_path()`.
Each USB device is uniquely identified by it's `path`, so multiple
lights of the same make and model can be opened successfully.

I don't yet have multiple instances of the same light to test with,
but this fix passes tests and appears to have not broken anything.