Luxafor Flag update
- added concrete reset method
- added color poperty to FlagState class
- added FlagState cmdbuf property for transient device commands
- update state property color members from cmdbuf colors
- updated docstrings

Not all Flag commands have a color (the red field is aliased to
another field in some cases), so I added a cmdbuf for tracking
the current command and track the color in the state property.
This will makes the is_on property easier to implement and be
less wrong.