Light Reorganization
Prior to the reorg, concrete implementations of a device resided
in a file named after the vendor e.g.: Blynclight class in the
embrava.py file. I decided to make packages for each vendor.
Each vendor package has a 'hardware.py' module with the hardware
specifics in it. :shrug: Maybe it's a good idea.