USBLight Update - expanded docstrings - added new property: strategy - changed `update` to use strategy property - added classmethod `all_lights` - converted `off` from an abstractmethod to regular method - added range checking to `blink` method's speed argument
Concrete subclasses no longer need to provide an `off` method, most
implementations call `on` with the color black so this was made
the default action. Subclasses may provide an `off` method that
updates the color property in addition to turning the light off.

Implementations of `on` and `blink` should call the superclass's
version prior to returning to the caller.

Most devices use the `hid.device.write` method to send data, however
at least one requries `hid.device.send_feature_report` instead. The
strategy property is used to communicate which method an
implementation requires. The default value is `hid.device.write`. This
keeps concrete implementations from having to provide their own
`update` method which may drift from the USBLight implementation (for
instance when locking was added).