Threading update
Normally, python threads run until they finish and are not
killable/cancellable.

The busylight.lights.thread.CancellableThread subclass accomplishes
this by requiring (hoping) the thread body is a generator function
that calls yield, giving the thread the opportunity to check if it's
been cancelled.

This update changes what is expected from the generator function
in an attempt to reduce code complexity and not result in weird
deadlocks. The generator should now yield a floating point value
which is the number of seconds that should pass before the next
pass thru the generator function. This simplifies the generator,
which was explicitly sleeping before or after yielding.

The only hardware light affected by this change is
busylight.lights.kuando.BusyLight which has a keepalive thread which
sends a keep alive packet to the light on a ~7 second cadence. I found
a bug where `light.release()` would not return due to an error in the
cancelling logic. With this fix a Kuando BusyLight can be released and
reaquired without hanging indefinitely.