USBLight Task Handling
The USBLight class grew instance methods for adding and
cancelling asyncio.Tasks associated with each light:
- property tasks: dict[str, asyncio.Task]
- add_tasks(self, name:str, awaitable:Awaitable)
- cancel_task(self, name:str)
- cancel_tasks(self)

The `tasks` property is a dictionary whose keys are string with
human-readable names for each task. The values are asyncio.Task
instances wrapping async functions.

When tasks are added, we check to see if a task exists with that
same name. If the task exists, we return it. Otherwise a new task
is created, recorded in the tasks property, and returned to the
caller.

Methods `cancel_task` and `cancel_tasks` are for cancelling tasks
either by name or all tasks associated with a USBLight instance.

The `reset` and `release` methods have been updated to cancel any
tasks current associated with the USBLight instance.

If a caller wants to completely quiesce a light, the `reset` method
is a better bet than calling the `off` method. The `off` methods
are abc.abstractmethods which generally do not cancel tasks and
only perform the off operation for the implementation. The `reset`
method is a USBLight method which will call `off` and then
cancel any tasks associated with the instance.