Update: busylight.api FastAPI overhaul
The busylight.api.BusylightAPI class was rewritten to replicate
most of the function of busylight.manager.LightManager. It was
easier for me to write this way, there may be an opportunity to
unify them as I start to grok asyncio better.

Several routes were removed:
- light/{id}/on/{color}
- light/{id}/blink/{color}
- light/{id}/blink/{color}/speed
- lights/on/{color}
- lights/blink/{color}
- lights/blink/{color}/speed
- lights/rainbow/{speed}

Update routes to take optional color and speed arguments:
- light/{id}/on?color={color}
- light/{id}/blink?color={color}&speed={speed}
- light/{id}/rainbow?speed={speed}
- lights/on?color={color}
- lights/blink?color={color}&speed={speed}

Added new routes:
- light/{id}/pulse?color={color}&speed={speed}
- light/{id}/fli?color_a={color}&color_b={color}&speed={speed}
- lights/pulse?color={color}&speed={speed}
- lights/fli?color_a={color}&color_b={color}&speed={speed}

Updated imports to reflect moved effects and color module.