Too many __version__ implementations, picked one.
Fixes issue #158 which was a result of trying to stuff a module
type into a string, specifically `busylight.__version__`. That
version was a duplicate of the code in `busylight.__init__` and
a residual from how versioning was handled in the project
Makefile. I think. I've slept since then. Keeping the __version__
in __init__ and removing __version__. Fixed the imports in
__main__, busylight_api and test_cli to reflect the new reality.