Fixed equality/less than to provide a stable sorted list of lights.
The Light path attribute is unique between instances, so using that
for the basis of sorting. For instances of HIDLight, the path
attribute was a bytes object rather than a string. Fixed that and
modified HIDLight.acquire to use "self.info['path']" for the bytes
representation it requires (rather than self.path).