Added first draft for a systemd-unitfile
I know, this file is not perfect, but it works.
I think the problem with the busyserve command is that it sends no "Success"-signal back to the main process, thats why I added the screen-command. I did this with many proprietary gameservers hosted on linux and thats fine, at least for me. Update / Improve if you wish.

The ExecStartPost can be customized to anyone's wishes of course, I added the retry and connect-timeout parameters because the Port is not available immediately at the time of EcexStartPost execution. I did not want to add a bash script (which sleeps a short time and then curls the server), but that were possible, too.