Fix for #416
An instance's id value was appended to a task name for reasons that
are lost in the mists of time. Presumably it was to disambiguate tasks
with the same name but different instances. However, this doesn't make
sense since tasks live in namespaces unique to instances, and the
disambiguation is not necessary.