Busy Tag File and Class Renaming
The previous file and class names included the string "SIA" which
is the Latvian translation of the phrase "Limited Liability Company".
Including that in the file and class names seems a little clunky.
I also shortened and title cased the vendor string to "Busy Tag"
to fit in with the other vendor strings when displayed via the
supported and available CLI commands.